server/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
